mental_health_bot/chatbot/requirements.txt
Django>=4.2
python-aiml>=0.9.1
scikit-learn>=1.4
joblib>=1.3
pandas>=2.2
numpy>=1.26